[digintlab-dep] Refactor into modules and add configurable primary-object type
Refactors the DigintLab DEP (Double Extortion) external-import connector
from a single ~680-line `src/main.py` into a `dep_connector/` package
and introduces a configurable `DEP_PRIMARY_OBJECT` (`report` / `incident`,
defaulting to `report`) so each announcement can be materialised as
either STIX shape while the surrounding identities / indicators /
relationships stay the same. Both primary-object ids are now derived
deterministically from the DEP `hashid` so later DEP updates that fill
in a previously-unknown victim merge into the existing announcement
instead of creating a new SDO.

Module split:

* `client_api.py` — Cognito auth + REST client, with explicit
  per-item `isinstance(item, dict)` validation on the JSON response so
  a future API change that returns an error object or a single-item
  dict surfaces as a clear `ValueError` at the client boundary rather
  than crashing downstream when `LeakRecord(**raw_item)` is fed a
  non-dict. Unused `import logging` and `logger = getLogger(...)` are
  removed to keep flake8 quiet.
* `config_loader.py` — YAML + env-var loader, honouring
  `OPENCTI_CONFIG_FILE`.
* `converter_to_stix.py` — the Pydantic `LeakRecord` model with field
  validators / normalisers, the `PrimaryObject` enum, and the
  `StixBuilder` covering `Identity` / `IntrusionSet` / `Location` /
  `Indicator` / `Report` / `Incident` / `Relationship`. `create_report`
  now derives its id from `uuid5(NAMESPACE_URL, f'dep-announcement:{hashid}')`
  matching the existing `create_incident` path — the previous
  `pycti.Report.generate_id(name, published)` keyed the id on
  `name` (built from `item.victim or item.victim_domain or "Unknown
  Victim"`) and `published`, so a later DEP update filling in a
  previously-unknown victim would produce a brand new Report SDO
  instead of upserting the existing announcement (and contradicting
  the README claim that primary-object ids are deterministic from
  `hashid`).
* `connector.py` — scheduling, state management, per-item routing.
  `_send_objects` now includes the `TLP_AMBER` marking-definition SDO
  in the bundle so `cleanup_inconsistent_bundle=True` can not strip
  every object's `object_marking_refs`; the deprecated `update=True`
  kwarg is dropped (removed from the helper's public contract under
  the `VC506` rule); per-cycle outcome is tracked through fetch
  failures and per-item failures so the Work is closed with
  `in_error=True` whenever anything went wrong instead of always
  reporting green-on-failure.

Signed-commit cleanup: the original two-commit history
(`fda59298`, `afce184a`) was authored by @notdodo but unverified
(the repo's `Check signed commits in PR` workflow rejects unsigned
commits). Squashed into a single GPG-signed commit with the review
fixes above folded in; original authorship is preserved via the

Local validation:

* `python -m py_compile` on every module — clean
* `black --check src/` — 6 files left unchanged
* `isort --profile black --check-only src/` — silent
* `flake8 --ignore=E,W src/` — silent

Closes #6510.
