+++ external-import/digintlab-dep/src/dep_connector/client_api.py
@@ -1,0 +1,109 @@
+import json
+from typing import TypeAlias
+
+import requests
+
+JsonPrimitive: TypeAlias = str | int | float | bool | None
+JsonValue: TypeAlias = JsonPrimitive | list["JsonValue"] | dict[str, "JsonValue"]
+DepApiItem: TypeAlias = dict[str, JsonValue]
+
+
+class DepClient:
+    def __init__(
+        self,
+        *,
+        login_endpoint: str,
+        api_endpoint: str,
+        api_key: str | None,
+        username: str | None,
+        password: str | None,
+        client_id: str,
+        dataset: str,
+        extended_results: bool,
+    ) -> None:
+        self.login_endpoint = login_endpoint
+        self.api_endpoint = api_endpoint
+        self.api_key = api_key
+        self.username = username
+        self.password = password
+        self.client_id = client_id
+        self.dataset = dataset
+        self.extended_results = extended_results
+
+    def authenticate(self) -> str:
+        headers = {
+            "Content-Type": "application/x-amz-json-1.1",
+            "X-Amz-Target": "AWSCognitoIdentityProviderService.InitiateAuth",
+        }
+        payload = {
+            "AuthParameters": {"USERNAME": self.username, "PASSWORD": self.password},
+            "AuthFlow": "USER_PASSWORD_AUTH",
+            "ClientId": self.client_id,
+        }
+        response = requests.post(
+            self.login_endpoint,
+            headers=headers,
+            json=payload,
+            timeout=30,
+        )
+        response.raise_for_status()
+        auth_payload: dict[str, dict[str, str]] = response.json()
+        token = auth_payload["AuthenticationResult"]["IdToken"]
+        if not token:
+            error = "Unable to retrieve IdToken from authentication response"
+            raise ValueError(error)
+        return token
+
+    def fetch_raw(
+        self,
+        start_date: str,
+        end_date: str,
+    ) -> list[DepApiItem]:
+        token = self.authenticate()
+        params: dict[str, str] = {
+            "ts": start_date,
+            "te": end_date,
+            "dset": self.dataset,
+            "full": "true",
+        }
+        if self.extended_results:
+            params["extended"] = "true"
+
+        headers = {
+            "X-Api-Key": self.api_key,
+            "Authorization": token,
+        }
+
+        response = requests.get(
+            self.api_endpoint,
+            headers=headers,
+            params=params,
+            timeout=60,
+        )
+        response.raise_for_status()
+        try:
+            payload = response.json()
+        except json.JSONDecodeError as exception:
+            message = "Unable to decode DEP API response"
+            raise ValueError(message) from exception
+
+        # The DEP announcements endpoint returns a JSON array of objects;
+        # validate the runtime shape explicitly so a future API change
+        # that returns an error object (``{"message": "..."}``) or a
+        # single-item dict surfaces as a clear ``ValueError`` at the
+        # client boundary instead of crashing downstream when
+        # ``LeakRecord(**raw_item)`` is fed a non-dict.
+        if not isinstance(payload, list):
+            error = (
+                "DEP API returned a non-list payload; "
+                f"expected list of items, got {type(payload).__name__}"
+            )
+            raise ValueError(error)
+        for index, item in enumerate(payload):
+            if not isinstance(item, dict):
+                error = (
+                    "DEP API returned a non-dict item at index "
+                    f"{index}: {type(item).__name__}"
+                )
+                raise ValueError(error)
+        return payload
